# backend/openapi.yaml
openapi: '3.0.3'
info:
  title: DetectiveML Backend API
  description: |
    Backend API for DetectiveML. All endpoints require Firebase Authentication — 
    clients send `Authorization: Bearer <firebaseIdToken>` to prove identity.
    The backend validates the token and performs Firebase operations using a service account.
    
    Gemini AI operations (chat, case generation, image generation, TTS) are proxied through
    the backend under `/api/gemini/`. The backend holds the GEMINI_API_KEY securely —
    the frontend never interacts with the Gemini SDK directly.
  version: '2.0.0'
servers:
  - url: http://localhost:4000
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID Token
      description: Firebase ID token obtained via `auth.currentUser.getIdToken()`
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      example:
        error: "Something went wrong."
...
    CaseData:
      type: object
      required: [id, title, type, description, difficulty, suspects, authorId]
      properties:
        id:
          type: string
          example: "case_abc123"
        title:
          type: string
          example: "The Missing Diamond"
        type:
          type: string
          example: "Larceny"
        description:
          type: string
          example: "A priceless diamond vanishes from a locked museum exhibit."
        difficulty:
          type: string
          enum: [Easy, Medium, Hard]
        isUploaded:
          type: boolean
          description: "Whether the case is published (read-only, controlled by publish endpoint)"
        authorId:
          type: string
          example: "uid_user123"
        authorDisplayName:
          type: string
          example: "Detective Jane"
        version:
          type: integer
          example: 3
        createdAt:
          type: number
          description: "Epoch timestamp (ms)"
        updatedAt:
          type: number
          description: "Epoch timestamp (ms)"
        heroImageUrl:
          type: string
        suspects:
          type: array
          items:
            type: object
        initialEvidence:
          type: array
          items:
            $ref: '#/components/schemas/Evidence'
        initialTimeline:
          type: array
          items:
            type: object
        officer:
          type: object
        partner:
          type: object
...
paths:
...
  /api/gemini/chat/suspect:
    post:
      summary: Get suspect interrogation response
      description: |
        Generates an AI response for a suspect during interrogation. Returns the response text,
        emotional state, aggravation change, revealed evidence, and timeline statements.
      tags: [Gemini AI]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [suspect, caseData, userInput, type, currentAggravation, isFirstTurn]
              properties:
                suspect:
                  type: object
                  description: The suspect being interrogated
                caseData:
                  $ref: '#/components/schemas/CaseData'
                userInput:
                  type: string
                  description: Player's question or action
                type:
                  type: string
                  enum: [talk, action]
                evidenceAttachment:
                  type: string
                  nullable: true
                  description: ID of evidence being presented
                currentAggravation:
                  type: number
                isFirstTurn:
                  type: boolean
                discoveredEvidence:
                  type: array
                  items:
                    type: object
                currentGameTime:
                  type: number
                conversationHistory:
                  type: array
                  items:
                    type: object
      responses:
        '200':
          description: AI suspect response
          content:
            application/json:
              schema:
                type: object
                properties:
                  text:
                    type: string
                  emotion:
                    type: string
                  aggravationDelta:
                    type: number
                  revealedEvidence:
                    type: array
                    items:
                      type: string
                  revealedTimelineStatements:
                    type: array
                    items:
                      type: object
                  hints:
                    type: array
                    items:
                      type: string
        '401':
          description: Unauthorized
        '500':
          description: AI generation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/gemini/chat/officer:
    post:
      summary: Get officer (Chief) chat response
      description: Returns AI dialogue from the officer/chief character when the player radios in.
      tags: [Gemini AI]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [caseData, userMessage, evidenceFound, notes, chatHistory]
              properties:
                caseData:
                  $ref: '#/components/schemas/CaseData'
                userMessage:
                  type: string
                evidenceFound:
                  type: array
                  items:
                    type: object
                notes:
                  type: object
                chatHistory:
                  type: object
      responses:
        '200':
          description: Officer response text
          content:
            application/json:
              schema:
                type: object
                properties:
                  text:
                    type: string
        '401':
          description: Unauthorized
        '500':
          description: AI generation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/gemini/chat/partner:
    post:
      summary: Get partner intervention response
      description: Generates a Good Cop, Bad Cop, examine, or hint intervention from the partner character.
      tags: [Gemini AI]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, suspect, caseData, history]
              properties:
                type:
                  type: string
                  enum: [goodCop, badCop, examine, hint]
                suspect:
                  type: object
                caseData:
                  $ref: '#/components/schemas/CaseData'
                history:
                  type: array
                  items:
                    type: object
                discoveredEvidence:
                  type: array
                  items:
                    type: object
      responses:
        '200':
          description: Partner intervention text
          content:
            application/json:
              schema:
                type: object
                properties:
                  text:
                    type: string
        '401':
          description: Unauthorized
        '500':
          description: AI generation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/gemini/chat/badcop-hint:
    post:
      summary: Get bad cop hint
      description: Generates a tactical hint based on unrevealed evidence after a bad cop intervention.
      tags: [Gemini AI]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [suspect, unrevealed, responseText]
              properties:
                suspect:
                  type: object
                unrevealed:
                  type: array
                  items:
                    type: object
                responseText:
                  type: string
      responses:
        '200':
          description: Hint text
          content:
            application/json:
              schema:
                type: object
                properties:
                  text:
                    type: string
        '401':
          description: Unauthorized
        '500':
          description: AI generation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/gemini/chat/case-summary:
    post:
      summary: Generate end-game case summary
      description: Generates a narrative summary of the completed case for the end-game screen.
      tags: [Gemini AI]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [caseData, gameResult, evidenceDiscovered]
              properties:
                caseData:
                  $ref: '#/components/schemas/CaseData'
                accusedId:
                  type: string
                  nullable: true
                gameResult:
                  type: string
                  enum: [SUCCESS, PARTIAL, FAILURE]
                evidenceDiscovered:
                  type: array
                  items:
                    type: object
      responses:
        '200':
          description: Summary text
          content:
            application/json:
              schema:
                type: object
                properties:
                  text:
                    type: string
        '401':
          description: Unauthorized
        '500':
          description: AI generation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/gemini/case/generate:
    post:
      summary: Generate a new case from a prompt
      description: |
        Uses AI to generate a complete mystery case from a user prompt.
        Returns full CaseData with suspects, evidence, timelines, alibis, and relationships.
        This is a long-running operation (may take 30-60 seconds).
      tags: [Gemini AI]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userPrompt:
                  type: string
                  description: Case theme/description (empty with isLucky for random)
                  example: "A murder at a jazz club in 1920s New York"
                isLucky:
                  type: boolean
                  description: If true and no prompt, generates a random theme
                  default: false
      responses:
        '200':
          description: Generated case data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseData'
        '401':
          description: Unauthorized
        '500':
          description: AI generation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/gemini/case/consistency:
    post:
      summary: Run AI consistency check on a case
      description: |
        Validates a case for internal consistency (timelines, relationships, evidence).
        Returns the updated case with fixes applied and a detailed report.
      tags: [Gemini AI]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [caseData]
              properties:
                caseData:
                  $ref: '#/components/schemas/CaseData'
                baseline:
                  $ref: '#/components/schemas/CaseData'
                editContext:
                  type: string
      responses:
        '200':
          description: Updated case and consistency report
          content:
            application/json:
              schema:
                type: object
                properties:
                  updatedCase:
                    $ref: '#/components/schemas/CaseData'
                  report:
                    type: object
        '401':
          description: Unauthorized
        '500':
          description: AI generation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/gemini/case/consistency/narrative:
    post:
      summary: Consistency check — narrative phase only
      description: |
        Runs the AI narrative consistency pass and merge. Does not regenerate images.
        Pair with POST /api/gemini/case/consistency/images using the returned updatedCase as mergedCase.
      tags: [Gemini AI]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [caseData]
              properties:
                caseData:
                  $ref: '#/components/schemas/CaseData'
                baseline:
                  $ref: '#/components/schemas/CaseData'
                editContext:
                  type: string
      responses:
        '200':
          description: Merged case and report (images unchanged)
          content:
            application/json:
              schema:
                type: object
                properties:
                  updatedCase:
                    $ref: '#/components/schemas/CaseData'
                  report:
                    type: object
        '401':
          description: Unauthorized
        '500':
          description: AI generation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/gemini/case/consistency/images:
    post:
      summary: Consistency check — image regeneration phase
      description: |
        Regenerates victim portraits, examination views, and evidence images after narrative merge.
        originalCaseData must be the case JSON from before the narrative pass (for invalidation diffs).
      tags: [Gemini AI]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mergedCase, originalCaseData]
              properties:
                mergedCase:
                  $ref: '#/components/schemas/CaseData'
                originalCaseData:
                  $ref: '#/components/schemas/CaseData'
      responses:
        '200':
          description: mergedCase mutated in place with new image URLs
          content:
            application/json:
              schema:
                type: object
                properties:
                  updatedCase:
                    $ref: '#/components/schemas/CaseData'
                  imagePipelineChanges:
                    type: array
                    description: Rows to merge into the consistency report (visual pass)
                    items:
                      type: object
                      properties:
                        description:
                          type: string
                        evidenceId:
                          type: string
        '400':
          description: Missing mergedCase or originalCaseData
        '401':
          description: Unauthorized
        '500':
          description: Image pipeline failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/gemini/case/edit:
    post:
      summary: Apply AI-driven edits to a case
      description: |
        Allows broad AI-driven editing of a case via natural language prompt.
        Handles theme changes, suspect modifications, evidence updates, etc.
      tags: [Gemini AI]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [caseData, userPrompt]
              properties:
                caseData:
                  $ref: '#/components/schemas/CaseData'
                userPrompt:
                  type: string
                  example: "Change the setting to a Victorian mansion"
                baseline:
                  $ref: '#/components/schemas/CaseData'
      responses:
        '200':
          description: Edited case and change report
          content:
            application/json:
              schema:
                type: object
                properties:
                  updatedCase:
                    $ref: '#/components/schemas/CaseData'
                  report:
                    type: object
        '401':
          description: Unauthorized
        '500':
          description: AI generation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/gemini/image/generate:
    post:
      summary: Generate a raw image
      description: Generates an image from a prompt using Gemini. Returns raw base64 PNG data.
      tags: [Gemini AI]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prompt]
              properties:
                prompt:
                  type: string
                aspectRatio:
                  type: string
                  default: "1:1"
                refImages:
                  type: array
                  items:
                    type: string
                mode:
                  type: string
                  enum: [create, edit, evidence]
                  default: create
                modelOverride:
                  type: string
      responses:
        '200':
          description: Base64 PNG image
          content:
            application/json:
              schema:
                type: object
                properties:
                  base64:
                    type: string
        '401':
          description: Unauthorized
        '500':
          description: Image generation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/gemini/image/evidence:
    post:
      summary: Generate an evidence image
      description: Generates a pixel-art evidence image and uploads it to Firebase Storage.
      tags: [Gemini AI]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [evidence, caseId, userId]
              properties:
                evidence:
                  type: object
                caseId:
                  type: string
                userId:
                  type: string
                refImage:
                  type: string
      responses:
        '200':
          description: Public URL of the uploaded evidence image
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
        '401':
          description: Unauthorized
        '500':
          description: Image generation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/gemini/image/variants:
    post:
      summary: Generate emotional portrait variants
      description: |
        Given a neutral base portrait, generates emotional variants (happy, angry, nervous, etc.)
        and uploads them to Firebase Storage.
      tags: [Gemini AI]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [neutralBase64, suspect, caseId, userId]
              properties:
                neutralBase64:
                  type: string
                  description: Base64 PNG of the neutral portrait
                suspect:
                  type: object
                caseId:
                  type: string
                userId:
                  type: string
      responses:
        '200':
          description: Map of emotion → Firebase Storage URL
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
        '401':
          description: Unauthorized
        '500':
          description: AI generation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/gemini/image/suspect-upload:
    post:
      summary: Generate suspect portraits from uploaded image
      description: |
        Takes a user-uploaded image and generates pixel-art portraits in the game style,
        then uploads all variants to Firebase Storage.
      tags: [Gemini AI]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [suspect, userImageBase64, caseId, userId]
              properties:
                suspect:
                  type: object
                userImageBase64:
                  type: string
                caseId:
                  type: string
                userId:
                  type: string
      responses:
        '200':
          description: Updated suspect with portrait URLs
          content:
            application/json:
              schema:
                type: object
        '401':
          description: Unauthorized
        '500':
          description: AI generation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/gemini/image/regenerate:
    post:
      summary: Regenerate a single suspect's portraits
      description: Regenerates all portraits (neutral + emotional variants) for a single suspect.
      tags: [Gemini AI]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [suspect, caseId, userId]
              properties:
                suspect:
                  type: object
                caseId:
                  type: string
                userId:
                  type: string
                theme:
                  type: string
                  default: "Noir"
      responses:
        '200':
          description: Updated suspect with new portrait URLs
          content:
            application/json:
              schema:
                type: object
        '401':
          description: Unauthorized
        '500':
          description: AI generation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/gemini/image/pregenerate:
    post:
      summary: Pre-generate all images for a case
      description: |
        Generates all portraits (suspects, officer, partner) and evidence images for a case.
        This is the longest-running operation — may take several minutes.
        Returns the full CaseData with all image URLs populated.
      tags: [Gemini AI]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [caseData, userId]
              properties:
                caseData:
                  $ref: '#/components/schemas/CaseData'
                userId:
                  type: string
      responses:
        '200':
          description: CaseData with all image URLs populated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseData'
        '401':
          description: Unauthorized
        '500':
          description: AI generation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/gemini/image/create:
    post:
      summary: Create an image from a user prompt
      description: Generates a pixel-art styled image from a freeform user prompt.
      tags: [Gemini AI]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userPrompt]
              properties:
                userPrompt:
                  type: string
                aspectRatio:
                  type: string
                  default: "3:4"
      responses:
        '200':
          description: Base64 PNG image
          content:
            application/json:
              schema:
                type: object
                properties:
                  base64:
                    type: string
        '401':
          description: Unauthorized
        '500':
          description: Image generation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...